refactor(server): move profileAliasMiddleware to its own file
Pair the middleware with its already-separate test file
(profile_alias_middleware_test.go). Pure move out of server_routes.go —
same package, no behavior change; the strings/logrus imports go with it.
